+++ src/main/java/net/silentchaos512/gear/item/gear/GearShovelItem.java
@@ -21,6 +21,7 @@ import net.minecraftforge.common.ToolActions;
 import net.silentchaos512.gear.api.item.GearType;
 import net.silentchaos512.gear.api.item.ICoreTool;
 import net.silentchaos512.gear.client.util.GearClientHelper;
+import net.silentchaos512.gear.config.Config;
 import net.silentchaos512.gear.util.GearData;
 import net.silentchaos512.gear.util.GearHelper;
 
@@ -42,8 +43,13 @@ public class GearShovelItem extends ShovelItem implements ICoreTool {
     private final GearType gearType;
 
     public GearShovelItem(GearType gearType) {
-        super(GearHelper.DUMMY_TIER, 0, 0f, GearHelper.getBaseItemProperties());
+        super(GearHelper.DEFAULT_DUMMY_TIER, 0, 0f, GearHelper.getBaseItemProperties());
         this.gearType = gearType;
+    }
+
+    @Override
+    public Tier getTier() {
+        return Config.Common.isLoaded() ? Config.Common.dummyToolTier.get() : GearHelper.DEFAULT_DUMMY_TIER;
     }
 
     @Override